rename files following CamelCase
